@@ -19,50 +19,48 @@ def generate_config(config_path: Path) -> None:
     :param config_path: path to config.
     """
     if config_path.exists():
-        print(f"Config {config_path} already exists")
+        logger.warning(f"Config {config_path} already exists")
         return
     Config().save(config_path)
-    print(f"Config successfully generated at {config_path}")
+    logger.info(f"Config successfully generated at {config_path}")
 
 
 def show_version() -> None:
     """Show installed version of a Music Background."""
     version = metadata.version("music_bg")
-    print(f"Music background v{version}")
+    logger.info(f"Music background v{version}")
 
 
-def print_processors(context: Context) -> None:  # noqa: WPS213, WPS210, WPS210, C901
+def print_processors(context: Context) -> None:
     """
     Print information about available image processors.
 
     :param context: music_bg context.
     """
-    print(" Processors ".center(80, "#"))
+    logger.info(" Processors ".center(80, "#"))
 
     for name, func in context.processors_map.items():
-        print("-" * 80)
-        print(f"name: {name}")
-        print("type: processor")
+        logger.info("-" * 80)
+        logger.info(f"name: {name}")
+        logger.info("type: processor")
         args = inspect.signature(func)
         custom_args = []
         for parameter in args.parameters.values():
             if parameter.annotation != Image:
                 custom_args.append(parameter)
         if custom_args:
-            print()
-            print(" args ".center(20, "="))
+            logger.info(" args ".center(20, "="))
             for arg in custom_args:
                 arg_info = f"* {arg.name}"
                 if arg.annotation != arg.empty:
                     arg_info = f"{arg_info}: {arg.annotation}"
                 if arg.default == arg.empty:
                     arg_info = f"{arg_info} (required)"
-                print(arg_info)
+                logger.info(arg_info)
         doc = inspect.getdoc(func)
         if doc is not None:
-            print()
-            print(" doc ".center(20, "="))
-            print(doc)
+            logger.info(" doc ".center(20, "="))
+            logger.info(doc)
 
 
 def print_variables(context: Context) -> None:
@@ -71,17 +69,16 @@ def print_variables(context: Context) -> None:
 
     :param context: current mbg context.
     """
-    print(" Variables ".center(80, "#"))
+    logger.info(" Variables ".center(80, "#"))
 
     for name, func in context.variables_providers.items():
-        print("-" * 80)
-        print(f"name: {name}")
-        print("type: variable")
+        logger.info("-" * 80)
+        logger.info(f"name: {name}")
+        logger.info("type: variable")
         doc = inspect.getdoc(func)
         if doc is not None:
-            print()
-            print(" doc ".center(20, "="))
-            print(doc)
+            logger.info(" doc ".center(20, "="))
+            logger.info(doc)
 
 
 def show_info(
@@ -101,10 +98,8 @@ def show_info(
     """
     show_version()
     if show_processors:
-        print()
         print_processors(context)
     if show_variables:
-        print()
         print_variables(context)
 
 
